Python/app/requirements.txt
langchain==0.0.76
openai
gTTS
ffmpeg-python
whisper
streamlit
pandas
numpy 
scipy
htbuilder
faiss-cpu
sentence-transformers
InstructEmbedding
googletrans==3.1.0a0